+++ src/handlers/narratable.rs
